refactor(runner): extract docker client from manager
Move all docker CLI invocation and JSON parsing into a new
dockerClient struct in docker_client.go, leaving manager.go focused
on container lifecycle orchestration.
